+++ lib/temporalio/connection.rb
@@ -1,6 +1,7 @@
 require 'temporalio/bridge'
 require 'temporalio/bridge/connect_options'
 require 'temporalio/bridge/tls_options'
+require 'temporalio/connection/tls_options'
 require 'temporalio/connection/test_service'
 require 'temporalio/connection/workflow_service'
 require 'temporalio/errors'
@@ -17,7 +18,7 @@ module Temporalio
     attr_reader :core_connection
 
     # @param address [String | nil] `host[:port]` for the Temporal server. Host defaults to `localhost:7233`.
-    # @param [TlsOptions | nil] tls
+    # @param [Temporalio::Connection::TlsOptions | nil] tls
     # @param [String | nil] identity
     # @param [Hash | nil] metadata
     def initialize(
@@ -41,7 +42,7 @@ module Temporalio
             client_private_key: tls.client_private_key,
             server_name_override: tls.server_name_override,
           ),
-          identity: identity || self.class.default_identity,
+          identity: identity || default_identity,
           metadata: metadata,
           retry_config: retry_config && Temporalio::Bridge::RetryConfig.new(
             initial_interval_millis: retry_config.initial_interval_millis,
@@ -82,55 +83,8 @@ module Temporalio
       [uri.host || 'localhost', uri.port || 7233]
     end
 
-    class << self
-      private
-
-      def default_identity
-        "#{Process.pid}@#{Socket.gethostname}"
-      end
-    end
-
-
-
-    class RetryConfig
-      # Initial backoff interval.
-      # @default 100
-      attr_reader :initial_interval_millis
-
-      # Randomization jitter to add.
-      # @default 0.2
-      attr_reader :randomization_factor
-
-      # Backoff multiplier.
-      # @default 1.5
-      attr_reader :multiplier
-
-      # Maximum backoff interval.
-      # @default 5000
-      attr_reader :max_interval_millis
-
-      # Maximum total time (optional).
-      attr_reader :max_elapsed_time_millis
-
-      # Maximum number of retries.
-      # @default 10
-      attr_reader :max_retries
-
-      def initialize(
-        initial_interval_millis: nil,
-        randomization_factor: nil,
-        multiplier: nil,
-        max_interval_millis: nil,
-        max_elapsed_time_millis: nil,
-        max_retries: nil
-      )
-        @initial_interval_millis = initial_interval_millis || 100
-        @randomization_factor = randomization_factor || 0.2
-        @multiplier = multiplier || 1.5
-        @max_interval_millis = max_interval_millis || 5_000
-        @max_elapsed_time_millis = max_elapsed_time_millis
-        @max_retries = max_retries || 10
-      end
+    def default_identity
+      "#{Process.pid}@#{Socket.gethostname}"
     end
   end
 end
