+++ sig/temporalio/connection/retry_config.rbs
@@ -1,0 +1,37 @@
+module Temporalio
+  class Connection
+    class RetryConfig
+      # Initial backoff interval.
+      # @default 100
+      attr_reader initial_interval_millis: Integer
+
+      # Randomization jitter to add.
+      # @default 0.2
+      attr_reader randomization_factor: Float
+
+      # Backoff multiplier.
+      # @default 1.5
+      attr_reader multiplier: Float
+
+      # Maximum backoff interval.
+      # @default 5000
+      attr_reader max_interval_millis: Integer
+
+      # Maximum total time (optional).
+      attr_reader max_elapsed_time_millis: Integer?
+
+      # Maximum number of retries.
+      # @default 10
+      attr_reader max_retries: Integer
+
+      def initialize: (
+          ?initial_interval_millis: Integer?,
+          ?randomization_factor: Float?,
+          ?multiplier: Float?,
+          ?max_interval_millis: Integer?,
+          ?max_elapsed_time_millis: Integer?,
+          ?max_retries: Integer?,
+        ) -> void
+    end
+  end
+end
